Move some backend code around
